Shard Client with Integration (#11)
Builds on https://github.com/xetdata/xet-core/pull/10.

Integrates the shard client with the rest of the framework.
The shard client registers all shards uploaded to CAS with the shard
server. Currently, the cas endpoint and the shard endpoint are
identical. If the CAS endpoint starts with local:// only, all the shard
server operations are skipped.

The file reconstruction queries are controlled by the
config.smudge_query_policy setting, which can be one of ServerOnly,
LocalOnly, or LocalFirst. If the CAS endpoint starts with local://, then
it is set to LocalOnly, otherwise it defaults to LocalFirst. It can also
be specified on the command line using e.g.
`--smudge_query_policy=local_only`.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>