mount watch import os unix only on non-windows (#38)
Automated windows build failed due to inability to import
`std::os::unix::fs::{MetadataExt, PermissionsExt}`.

This PR changes that import to be conditional on target os not being
windows. Confirmed on windows self-hosted runner machine that
compilation is successful and can clone a repo with compiled debug
binary with this change.