Allow xet cli /pyxet to work with incomplete git config (no email, or username) (#40)
We are much too aggressive about checking user.email and user.name.
Those are only necessary if you need to write to a git repo. For pyxet /
xetcli, we only need ever need to pull. Hence it is optional and not
required.
The only places where we really need to check this is in `git xet mount`
and `git xet clone` which already checks it explicitly.