Potential fix for wrong Python binary detection
It is difficult to reliably identify the location of the python
binary which was used to run the 'xet' command. This is required
to "relaunch" xet as a daemon. Here, instead we directly relaunch
the 'xet' command directly, hoping that the system will "do the right
thing" (i.e. exec the script with the shebang command at the start of
'xet').