begin_write_transaction now lets the synchronize step be optional. (#91)
While we're still figuring this out, the "Synchronize with remote" step
is now enabled only if the environment variable XET_FETCH_ALL_SHARDS is
set. If this is not set, then this step is skipped, and all dedup relies
on the shard hints mechanism. However, this may have unforeseen
consequences, so for now it may still be enabled for debugging and
profiling.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>