background puts support for CASClient, implemented in PassthroughStagingClient (#98)
This PR changes a fundamental interface component of the CAS Client. It
allows the put() method to be backgrounded. The background put()
capability has been implemented in the PassthroughStagingClient (at the
moment that is the most convenient location to enable the xetblob and
hence Python API to get parallel uploads).

This means that for corectness, all puts() must be followed by a
flush(). I fixed all the locations I can find.