Updated the current progress bar utitilty to avoid &mut self
This PR updates the current progress bar used by the git_stream
functions and CAS uploading routines to avoid &mut self, thus
allowing it to be easily wrapped by pyxet and passed between python and
rust.  This has the additional advantage of avoiding a mutex, as was the case in
git_stream.

The base functionality and format has not changed.

To test out the different formats and use cases, run the produced
executable test_progress_reporting.