Logging File improvements. (#155)
Currently, the log files are written to a fixed path. However, with this
configuration, it means that the operation logs saved in the integration
tests overwrite previous operations, making debugging things with
multiple invocations of git difficult. This PR adds "{timestamp}" and
"{pid}" fields to the log path, which allows XET_LOG_PATH to be generate
unique file paths for each process log.

For example,

```
export XET_LOG_PATH="$HOME/logs/log_{timestamp}_{pid}.txt"
```

could result in the log file 

```
logs/log_2023-10-19T15-28-14.580960+00-00_71376.txt
```

In addition, logging now ensures that the directory the log would be
written to exists, and that relative paths work correctly.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>