Local config file to enable github forks. (#147)
This PR enables github integration that works through forks and pull
requests using an additional file in .xet/config.toml. This file that
gives the url of the repository where Xet is initialized. Using this, on
init, the filter process probes that repository to fetch the notes for
the salt and the merkledb, if they exist. This file looks like:

```
[upstream]
origin_type = "github"
user_name = "<repo user name>"
repo_name = "<repo name>"
```

This file, if present on HEAD, is picked up by the configuration.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>