Ensure git will not block on waiting for user input. (#161)
With integrating git with different endpoints, git will sometimes block
on waiting for user input when the authentication configurations are not
correct. Since the integration methods probe several connection methods
(https, ssh), this causes an issue. This PR prevents this, and the git
process will error out in such a case.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>