Clear Initialization logic + More Init options (#169)
To enable the user integration scenarios, the init logic needed to be
modified. However, there was quite a bit of logic that had become very
convoluted, making it difficult to properly trace what was going on. In
addition, to test these scenarios, we needed to add logic to the
repository init fields to set up origin repos for these scenarios
correctly.

This PR organizes the init logic into two functions,
perform_explicit_setup and perform_implicit_setup, and then moves the
actions involved in these two into separate and clearly described linear
tasks. Overall, the number of various aggregate setup functions is
reduced from 5-6 down to these two, making the logic much easier to
follow.

In addition, two options are added to the explicit setup function,
namely a minimal flag, which only writes the basic commit files, and a
an option to include the .xet/config.toml file as part of the initial
commit. These are needed for testing purposes.
Furthermore, logic was added so that init can now work more fully on
bare repos as well.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>