Full tests for integration scenarios (#160)
This PR adds tests for all the integration scenarios. In this scenario,
it is assumed that the only things that propagate through events like
forks, PR merges, and the like are files that are checked in to the
repo. Here, we have tests that cover forking, merging, pushes, pulls,
etc. to ensure that all notes are properly handled and all files can
always be reconstructed properly.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>