Fix for intermediate test failure. (#189)
git commit seems to barf out an exit code 1 sometimes if it thinks that
the file hasn't changed from the index based on timestamps. This can
sometimes causes the test_github_integration_simple test to
intermittently fail depending on where the boundary of the 1 second tick
used for difference checking. This PR fixes this.

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>