Avoided loading file in cache to extract block id. (#203)
In our cache, the key is encoded in the filename and comprised of the
CAS name, the block index, and the block size. However, on loading the
cache, this information is read from the file instead of extracted from
the file name. On my local cache of roughly 800 files, this cuts the
init time for the CAS cache from 1.5 seconds to ~100 ms.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>