Fix for build breakage (#221)
It turns out there is a weird thing in rust's proc_macro_system that can
cause the previous commit (incorporating a source drop of thiserror) to
break the build.

In this case, registering a new version of thiserror may cause, on
compilation, other packages (in this case opentelemetry) to use our
source-dropped version of thiserror instead of the full package. This
breaks the build as our version adds a dependency on the tracing
package, but this dependency is only propagated through the proc macro
and doesn't seem -- for whatever reason -- to cause the compiler to make
opentelemetry also depend on the tracing package. This breaks the build,
as the proc macro generates code that then has a dependency on tracing,
which opentelemetry doesn't depend on.

This PR fixes this by:
1. Renaming the local source drop to xet_error. This way there won't be
any confusion.
1. Moving the dependency on tracing out of the package. This is done by
changing the enabling code to register a callback instead of setting a
flag. The local callback then implements the tracing functionality.

If you've encountered the above error, update to this PR. If your build
breaks, you need to really nuke things.

1. Remove everything in the ~/.cargo/registry/ directory.
2. cargo clean
3. rebuild.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>