Compact in-memory lookup table for chunks (#218)
- Implements a compact, in-memory hash table that takes a partial chunk
hash and indicates which shard that chunk would be part of and where it
is located in the shard.

- The hash table is simply an index of a 64 bit truncated hash to a 16
bit index of the shard, along with 32 + 16 bits for the cas block
location and offset. This allows the lookup process to skip the linear
search between shards and the bisection search within the shard.

- The number of chunks loaded is limited by XET_MAX_CHUNK_CACHE_SIZE,
which defaults to 64M entries. Each element takes 16 bytes (plus the
size overhead of the hash table). The chunks are prioritized by most
recent shards first.

- On a benchmark repo with a collection of 14 shards and about 10M
chunks, this increased the cleaning filter process from 70-90 MB/s to
300-450 MB/s.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>