Adding size threshold as a config (#227)
Changes the small file threshold used in the pointer file translators
from a constant to a config (e.g. `XET_CAS_SIZETHRESHOLD`). This allows
applications like xet sync to treat all uploaded files as pointer files
(by setting the threshold to 0) so that the sync doesn't overwhelm
XetHub when uploading a dataset of thousands of moderately sized
(~100KB) text files.

We limit this config to be in the range: `[0, SMALL_FILE_THRESHOLD]` so
that users don't try to upload files larger than our current
pointer-file threshold.