Background async iterator for parallel processing of streams and chunking.  (#224)
This PR splits reading, chunking, and dedup into three separate threads,
causing a 2-3X speedup for the clean filter. Benchmarks now show 500-600
MB/s on larger files on an SSD.

This PR implements several utilities to make this process testable and
simple.

trait AsyncIterator<ErrorType>: An async iterator type that is templated
over the error type and Item to replace the very specific AsyncIterator
in merkledb. This is now defined in parutils. The async_chunk_iterator
was then modified to take and conform to this iterator trait.

AsyncDataIterator trait -- Simply a specialization of the AsyncIterator
trait to use ErrorType = GitXetRepoError and Item = Vec<u8>. This is now
the type used throughout the data_processing_* classes. In addition, the
existing AsyncFileIterator now conforms to this trait.

BufferedAsyncIterator: An async iterator implementation that is
initialized using any AsyncIterator-conforming object and a buffer size
and spawns a background tokio thread to keep the iterator buffer full.
At the end, the original iterator can be retrieved using the finish()
method. Additionally, this iterator can be initialized using a
collection of starting objects, meaning that we don't need
AsyncIteratorWithPutback any more.

Now, with these utilities, the code to background both the data loading
and the chunking is quite simple:

```
        // load enough data from reader to determine whether this file is just pass-through or not.
        let starting_data = { ... } 
        
        // Test if the file is pass through...  If it is return. 
        ...
        
        // Now, keep loading the data on a background thread.
        let raw_data_iter =
            BufferedAsyncIterator::new_with_starting_data(starting_data, reader, Some(512)).await?;

        // Pass this into the chunker to chunk on another background thread.
        let mut generator =
            BufferedAsyncIterator::new(async_chunk_target_default(raw_data_iter), Some(4096))
                .await?;
 ```

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>