Updated logic around user name and user email (#242)
This PR updates the logic around handling the user.name and user.email
fields in the config. These were needed for adding notes, as the notes
needed a signature field to be filled out. However, as these fields are
rarely visible, we now use the following logic:
- If user.name and user.email are available from git, use those.   
- Otherwise, if the user name and email are available from the xet
config, use those.
- Finally, if none of these are available, use "Xet System" for the name
and "system@xethub.com" for the backup values.

Now, with this change, there are no longer any checks for these fields.
Tests are updated to ensure they work in this environment.

Resolves https://github.com/xetdata/xethub/issues/5021,
https://github.com/xetdata/xethub/issues/4718,
https://github.com/xetdata/xethub/issues/2561.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>