Tableau Workbook Summary (#276)
Adds Tableau Workbook Summarization to git-xet, allowing us to summarize
`.twb` files.
The summarization works is done in the following high-level steps:
1. As blocks are cleaned from the `.twb` file, they're added to a buffer
by the `TwbAnalyzer`
2. Once complete, the `TwbAnalyzer::finalize()` method is called, which
kicks off the summarization.
3. The buffered twb file is converted to a raw XML tree (using roxmltree
crate).
4. The XML node tree is then converted into a `TwbRaw` struct,
containing a series of "raw" structs that roughly map to the XML nodes
in the file.
5. We then process the `TwbRaw` struct into a `TwbSummary` struct, which
provides the semantic concepts that the XetHub UI / Users expect (e.g.
logical tables and display names).
6. This `TwbSummary` is then returned back to the filter logic, where it
can be serialized and stored in the summary_db for the repo.

The layout of the `tableau_summary` crate follows fairly closely to the
above structure with a module for the twb summarization and sub-modules
for the `raw` and `summary` passes.

Note: there is likely a more optimal processing of the file in a
streaming manner, however, this is much more complex than simply going
through 3 fairly isolated passes. If performance becomes a problem, we
can optimize later.