Tableau Summary Refactoring (#343)
Fixes: TAB-83

Refactors the TwbSummary and TdsSummary to use a versioned enum for the
summary. This will allow us to add new schemas for the summaries as
needed while remaining backwards compatible with older summaries.

Details of testing can be found in:
https://www.notion.so/xethub/Summary-Versioning-Compatibility-21f173305a9c44dea05c33ed5b176546.

Note that old clients won't be able to parse newer schemas (e.g. V2+).