Change zones to be allow multiple top-level Zone objects (#362)
Fixes: TAB-115

Since the `<zones>` tag might have multiple `<zone>` tags within, we
need to change the Raw and Summary Dashboards to support this.

Note: this bumps up the schema version of the TwbSummary to 3 since this
is a type change. We should make sure that the UI is able to accept
these V3 schemas before merging.