Drop some info statements to debug. (#361)
info! is too verbose for a lot of things that have been stable for a
while, so this drops many of these to debug to make info! more
informative. This also adds some summary info! commands that pull
together larger blocks of statements with more aggregation.
Also, for tracking transactions, this PR adds a tag field so that
reporting on transactions can be traced.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>