repo migrate:  Push tags in staged manner just like branches (#397)
git xet repo migrate uses a backoff with branches to avoid pushing
hundreds of branches at once, but it does not do so with tags. Tags are
treated the same way as branches, however, so a repo with a lot of tags
can still lead to timeouts.

This PR fixes this by pushing all references in chunks. As such, these
references can all be pushed using the same mechanism.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>