Cap xetblob write chunk size (#399)
Fix CLI-190.

Let xetblob write chunks at the same size as git so as not to blow up
underlying clean process memory usage. Note that this however doesn't
handle memory usage intelligently when multiple files are cleaned
simultaneously -- that will be left to the users to tune the degree of
parallelism.