More fixes for note conversion path in repo migration.  (#396)
Currently, the note conversion ran as a separate stage after the regular
conversion, in which all blob oids in that stage reachable by notes and
their dependents were considered note blobs. The exception to the
building of this dependency tree is that attached commits that were
already converted would be ignored.

The problem is that many blobs in this path may not actually be note
blobs and should be put through the filter. The cases missed:
- notes attached to blobs. 
- dependents of notes attached to trees. (notes have to be attached to a
previously converted commit to turn off the switch saying it's a note).

This approach tracks whether an oid can be considered as a note through
the different stages so blobs are filtered and not filtered properly.

---------

Co-authored-by: Hoyt Koepke <hoytak@xethub.com>